AVX-60903 Addressed comments
