+++ aviatrix/provider.go
@@ -111,7 +111,7 @@ func Provider() *schema.Provider {
 			"aviatrix_distributed_firewalling_intra_vpc":                      resourceAviatrixDistributedFirewallingIntraVpc(),
 			"aviatrix_distributed_firewalling_origin_cert_enforcement_config": resourceAviatrixDistributedFirewallingOriginCertEnforcementConfig(),
 			"aviatrix_distributed_firewalling_policy_list":                    resourceAviatrixDistributedFirewallingPolicyList(),
-			"aviatrix_distributed_firewalling_zero_trust_rule":                resourceAviatrixDistributedFirewallingZeroTrustRule(),
+			"aviatrix_distributed_firewalling_default_action_rule":            resourceAviatrixDistributedFirewallingDefaultActionPolicy(),
 			"aviatrix_distributed_firewalling_proxy_ca_config":                resourceAviatrixDistributedFirewallingProxyCaConfig(),
 			"aviatrix_edge_csp":                                               resourceAviatrixEdgeCSP(),
 			"aviatrix_edge_csp_ha":                                            resourceAviatrixEdgeCSPHa(),
