+++ aviatrix/resource_aviatrix_distributed_Firewalling_default_action_policy_test.go
@@ -1,0 +1,104 @@
+package aviatrix
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"os"
+	"strings"
+	"testing"
+
+	"github.com/AviatrixSystems/terraform-provider-aviatrix/v3/goaviatrix"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
+)
+
+func TestAccAviatrixDistributedFirewallingDefaultActionPolicy_basic(t *testing.T) {
+	skipAcc := os.Getenv("SKIP_DISTRIBUTED_FIREWALLING_DEFAULT_ACTION_POLICY")
+	if skipAcc == "yes" {
+		t.Skip("Skipping Distributed Firewalling Default Action Policy test as SKIP_DISTRIBUTED_FIREWALLING_DEFAULT_ACTION_POLICY is set")
+	}
+	resourceName := "aviatrix_distributed_firewalling_default_action_policy.test"
+
+	resource.Test(t, resource.TestCase{
+		PreCheck: func() {
+			testAccPreCheck(t)
+		},
+		Providers:    testAccProvidersVersionValidation,
+		CheckDestroy: testAccDistributedFirewallingDefaultActionPolicyDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccDistributedFirewallingDefaultActionPolicyBasic(),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckDistributedFirewallingDefaultActionPolicyExists(resourceName),
+					resource.TestCheckResourceAttr(resourceName, "action", "PERMIT"),
+					resource.TestCheckResourceAttr(resourceName, "logging", "true"),
+				),
+			},
+			{
+				ResourceName:      resourceName,
+				ImportState:       true,
+				ImportStateVerify: true,
+			},
+		},
+	})
+}
+
+func testAccDistributedFirewallingDefaultActionPolicyBasic() string {
+	return `
+resource "aviatrix_distributed_firewalling_default_action_policy" "test" {
+    action  = "PERMIT"
+    logging = true
+}
+`
+}
+
+func testAccCheckDistributedFirewallingDefaultActionPolicyExists(n string) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+		rs, ok := s.RootModule().Resources[n]
+		if !ok {
+			return fmt.Errorf("no Distributed Firewalling Default Action Policy resource found: %s", n)
+		}
+		if rs.Primary.ID == "" {
+			return fmt.Errorf("no Distributed Firewalling Default Action Policy ID is set")
+		}
+
+		meta := testAccProviderVersionValidation.Meta()
+		client, ok := meta.(*goaviatrix.Client)
+		if !ok {
+			return fmt.Errorf("failed to assert meta as *goaviatrix.Client")
+		}
+
+		_, err := client.GetDistributedFirewallingDefaultActionPolicy(context.Background())
+		if err != nil {
+			return fmt.Errorf("failed to get Distributed Firewalling Default Action Policy status: %w", err)
+		}
+
+		if strings.Replace(client.ControllerIP, ".", "-", -1) != rs.Primary.ID {
+			return fmt.Errorf("distributed firewalling default action policy ID not found")
+		}
+
+		return nil
+	}
+}
+
+func testAccDistributedFirewallingDefaultActionPolicyDestroy(s *terraform.State) error {
+	meta := testAccProviderVersionValidation.Meta()
+	client, ok := meta.(*goaviatrix.Client)
+	if !ok {
+		return fmt.Errorf("failed to assert meta as *goaviatrix.Client")
+	}
+
+	for _, rs := range s.RootModule().Resources {
+		if rs.Type != "aviatrix_distributed_firewalling_default_action_policy" {
+			continue
+		}
+
+		_, err := client.GetDistributedFirewallingDefaultActionPolicy(context.Background())
+		if err == nil || !errors.Is(err, goaviatrix.ErrNotFound) {
+			return fmt.Errorf("distributed firewalling default action policy configured when it should be destroyed")
+		}
+	}
+
+	return nil
+}
